refactor(editor): reorganize standalone plugin files
