+++ experiments/generic-editor/src/plugins/standalone/components/StandaloneLayout.jsx
